+++ src/main/scala/com/convergencelabs/convergence/server/datastore/domain/ModelOperationStore.scala
@@ -135,7 +135,7 @@ class ModelOperationStore private[domain](dbProvider: DatabaseProvider)
       |  version >= :version
       |ORDER BY version ASC""".stripMargin
 
-  def getOperationsAfterVersion(id: String, version: Long, limit: Option[Int] = None): Try[List[ModelOperation]] = withDb { db =>
+  def getOperationsAfterVersion(id: String, version: Long, limit: Option[Long] = None): Try[List[ModelOperation]] = withDb { db =>
     val query = OrientDBUtil.buildPagedQuery(GetOperationsAfterVersionQuery, QueryLimit(limit), QueryOffset())
     val params = Map(Constants.ModelId -> id, Fields.Version -> version)
     OrientDBUtil
